+++ .gitignore
@@ -1,2 +1,2 @@
 coverage.txt
-/cm
+/cmd/carmirror
