+++ Makefile
@@ -11,7 +11,7 @@ clean:
 
 build:
 	go build ./...
-	go build -o ./cm ./plugin/cli/carmirror.go
+	go build -o ./cmd/carmirror ./cmd/carmirror.go
 
 setup-plugin:
 	grep -v carmirror ../kubo/plugin/loader/preload_list > ../kubo/plugin/loader/preload_list.tmp
@@ -30,25 +30,25 @@ test:
 	go test ./... -v --coverprofile=coverage.txt --covermode=atomic
 
 sharness:
-	cp test/sharness/$(TEST_NAME).sh ../kubo/test/sharness/ && cp -R test/sharness/$(TEST_NAME)-data ../kubo/test/sharness/ && cp cm ../kubo/test/bin/cm
+	cp test/sharness/$(TEST_NAME).sh ../kubo/test/sharness/ && cp -R test/sharness/$(TEST_NAME)-data ../kubo/test/sharness/ && cp ./cmd/carmirror ../kubo/test/bin/carmirror
 	$(MAKE) -C ../kubo/test/sharness $(TEST_NAME).sh
-	rm -rf ../kubo/test/sharness/$(TEST_NAME).sh ../kubo/test/sharness/$(TEST_NAME)-data ../kubo/test/bin/cm
+	rm -rf ../kubo/test/sharness/$(TEST_NAME).sh ../kubo/test/sharness/$(TEST_NAME)-data ../kubo/test/bin/carmirror
 
 sharness-v:
-	cp test/sharness/$(TEST_NAME).sh ../kubo/test/sharness/ && cp -R test/sharness/$(TEST_NAME)-data ../kubo/test/sharness/ && cp cm ../kubo/test/bin/cm
+	cp test/sharness/$(TEST_NAME).sh ../kubo/test/sharness/ && cp -R test/sharness/$(TEST_NAME)-data ../kubo/test/sharness/ && cp ./cmd/carmirror ../kubo/test/bin/carmirror
 	$(MAKE) -C ../kubo/test/sharness deps
 	cd ../kubo/test/sharness && ./$(TEST_NAME).sh -v
-	rm -rf ../kubo/test/sharness/$(TEST_NAME).sh ../kubo/test/sharness/$(TEST_NAME)-data ../kubo/test/bin/cm
+	rm -rf ../kubo/test/sharness/$(TEST_NAME).sh ../kubo/test/sharness/$(TEST_NAME)-data ../kubo/test/bin/carmirror
 
 sharness-no-deps:
-	cp test/sharness/$(TEST_NAME).sh ../kubo/test/sharness/ && cp -R test/sharness/$(TEST_NAME)-data ../kubo/test/sharness/ && cp cm ../kubo/test/bin/cm
+	cp test/sharness/$(TEST_NAME).sh ../kubo/test/sharness/ && cp -R test/sharness/$(TEST_NAME)-data ../kubo/test/sharness/ && cp ./cmd/carmirror ../kubo/test/bin/carmirror
 	cd ../kubo/test/sharness && ./$(TEST_NAME).sh
-	rm -rf ../kubo/test/sharness/$(TEST_NAME).sh ../kubo/test/sharness/$(TEST_NAME)-data ../kubo/test/bin/cm
+	rm -rf ../kubo/test/sharness/$(TEST_NAME).sh ../kubo/test/sharness/$(TEST_NAME)-data ../kubo/test/bin/carmirror
 
 sharness-no-deps-v:
-	cp test/sharness/$(TEST_NAME).sh ../kubo/test/sharness/ && cp -R test/sharness/$(TEST_NAME)-data ../kubo/test/sharness/ && cp cm ../kubo/test/bin/cm
+	cp test/sharness/$(TEST_NAME).sh ../kubo/test/sharness/ && cp -R test/sharness/$(TEST_NAME)-data ../kubo/test/sharness/ && cp ./cmd/carmirror ../kubo/test/bin/carmirror
 	cd ../kubo/test/sharness && ./$(TEST_NAME).sh -v
-	rm -rf ../kubo/test/sharness/$(TEST_NAME).sh ../kubo/test/sharness/$(TEST_NAME)-data ../kubo/test/bin/cm
+	rm -rf ../kubo/test/sharness/$(TEST_NAME).sh ../kubo/test/sharness/$(TEST_NAME)-data ../kubo/test/bin/carmirror
 
 update-deps:
 	go get -u ./... & go mod tidy
